ai-assistant-bot/requirements.txt
discord.py==2.3.2
anthropic==0.47.0
python-dotenv==1.0.0
aiohttp==3.9.5
beautifulsoup4==4.12.3
